% Auto-generated: do not edit by hand
\name{functionalComponent}

\alias{functionalComponent}

\title{FunctionalComponent component}

\description{
Some arbitrary description
}

\usage{
functionalComponent(id=NULL, value=NULL, n_clicks=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}

\item{value}{The value displayed in the input.}

\item{n_clicks}{Number of time button was clicked}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

